fix: fix wrong route path of home header input and rename to feed
