+++ apps/web/src/app/features/feed/pages/index.ts
@@ -1,6 +1,6 @@
 export * from './HomePage';
 export * from './FeedPage';
 export * from './SearchPage';
-export * from './PostDetailPage';
+export * from './FeedDetailPage';
 export * from './CreateFeedPage';
-export * from './UpdatePostPage';
+export * from './UpdateFeedPage';
